+++ app/resources/views/pages/dominion/magic.blade.php
@@ -291,6 +291,50 @@
         </div>
 
     </div>
+    <div class="row">
+
+        <div class="col-md-12 col-md-9">
+        <div class="box box-primary">
+            <div class="box-header with-border">
+                <h3 class="box-title"><i class="ra ra-burning-embers"></i> Spells affecting your dominion</h3>
+            </div>
+            <div class="box-body table-responsive no-padding">
+                <table class="table table-hover">
+                    <colgroup>
+                        <col width="150">
+                        <col>
+                        <col width="100">
+                        <col width="200">
+                    </colgroup>
+                    <thead>
+                        <tr>
+                            <th>Spell</th>
+                            <th>Effect</th>
+                            <th class="text-center">Duration</th>
+                            <th class="text-center">Cast By</th>
+                        </tr>
+                    </thead>
+                    <tbody>
+                        @foreach ($spellCalculator->getActiveSpells($selectedDominion) as $activeSpell)
+                            <tr>
+                                <td>{{ $activeSpell->spell->name }}</td>
+                                <td>{{ $spellHelper->getSpellDescription($activeSpell->spell) }}</td>
+                                <td class="text-center">{{ $activeSpell->duration }}</td>
+                                <td class="text-center">
+                                    @if ($activeSpell->cast_by_dominion_id == $selectedDominion->id || $selectedDominion->realm->dominions->pluck('id')->contains($activeSpell->cast_by_dominion_id) || $target->getSpellPerkValue('surreal_perception'))
+                                        <a href="{{ route('dominion.realm', $activeSpell->castByDominion->realm->number) }}">{{ $activeSpell->castByDominion->name }} (#{{ $activeSpell->castByDominion->realm->number }})</a>
+                                    @else
+                                        <em>Unknown</em>
+                                    @endif
+                                </td>
+                            </tr>
+                        @endforeach
+                    </tbody>
+                </table>
+            </div>
+        </div>
+
+    </div>
 @endsection
 
 @push('page-styles')